requirement.txt
sentence-transformers
pandas
ipykernel
Flask 
werkzeug
google-cloud-vision
wordsegment
-e .